pyproject.toml
[project]
name = "ai-platform"
version = "0.1.0"
description = "Universal Agent Platform - Profile 기반 도메인별 AI 에이전트 플랫폼"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "langgraph>=0.3.0",
    "langchain-core>=0.3.0",
    "httpx>=0.28.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "numpy>=1.26.0",
    "tiktoken>=0.8.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0",
]
local = [
    "sentence-transformers>=3.0.0",
    "torch>=2.0.0",
    "transformers>=4.40.0",
]
openai = [
    "openai>=1.50.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["src"]
